+++ .idea/misc.xml
@@ -4,8 +4,9 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/GeekBrains/Calculator2/app/src/main/res/drawable/ic_launcher_background.xml" value="0.238" />
-        <entry key="..\:/GeekBrains/Calculator2/app/src/main/res/layout-land/activity_main.xml" value="0.1" />
-        <entry key="..\:/GeekBrains/Calculator2/app/src/main/res/layout/activity_main.xml" value="0.1" />
+        <entry key="..\:/GeekBrains/Calculator2/app/src/main/res/layout-land/activity_main.xml" value="0.16" />
+        <entry key="..\:/GeekBrains/Calculator2/app/src/main/res/layout/activity_main.xml" value="0.16" />
+        <entry key="..\:/GeekBrains/Calculator2/app/src/main/res/layout/activity_settings.xml" value="0.2" />
       </map>
     </option>
   </component>
