+++ app/src/main/res/layout-land/activity_main.xml
@@ -5,28 +5,16 @@
     android:layout_height="match_parent"
     android:background="@drawable/landscape_rock_sky_road_texture_old_931138_pxhere_com__1_"
     tools:context=".MainActivity">
-    <RadioGroup
-        android:id="@+id/radioButtons"
-        android:layout_width="match_parent"
+
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/button_settings"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        >
+        android:text="@string/btn_settings"
+        android:textSize="@dimen/button_settings"
+        tools:ignore="MissingConstraints"
+        android:alpha="@dimen/alpha"/>
 
-
-        <com.google.android.material.radiobutton.MaterialRadioButton
-            android:id="@+id/radioButtonThemeCalculator"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/theme_calculator"
-            android:textColor="@color/white" />
-
-
-        <com.google.android.material.radiobutton.MaterialRadioButton
-            android:id="@+id/radioButtonMyThemeGrey"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/my_theme_grey"
-            android:textColor="@color/white" />
-    </RadioGroup>
 
 
     <GridLayout
