+++ app/src/main/res/layout/activity_main.xml
@@ -6,28 +6,13 @@
     android:background="@drawable/landscape_rock_sky_road_texture_old_931138_pxhere_com"
     tools:context=".MainActivity">
 
-    <RadioGroup
-        android:id="@+id/radioButtons"
-        android:layout_width="match_parent"
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/button_settings"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-       >
-
-
-    <com.google.android.material.radiobutton.MaterialRadioButton
-        android:id="@+id/radioButtonThemeCalculator"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/theme_calculator"
-        android:textColor="@color/white" />
-
-
-    <com.google.android.material.radiobutton.MaterialRadioButton
-        android:id="@+id/radioButtonMyThemeGrey"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/my_theme_grey"
-        android:textColor="@color/white" />
-    </RadioGroup>
+        android:text="@string/btn_settings"
+        android:textSize="@dimen/button_settings"
+        tools:ignore="MissingConstraints" />
 
     <GridLayout
         android:layout_width="wrap_content"
