+++ app/src/main/res/values/colors.xml
@@ -8,4 +8,6 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="grey">#828385</color>
+    <color name="settings">#535863</color>
+
 </resources>
