+++ app/src/main/res/values/strings.xml
@@ -20,6 +20,9 @@
     <string name="btn_plus_minus">+/-</string>
     <string name="btn_percentage">%</string>
     <string name="result_text">0</string>
-    <string name="theme_calculator">Black</string>
+    <string name="my_theme_black">Black</string>
     <string name="my_theme_grey">Gray</string>
+    <string name="btn_settings">Настройки темы</string>
+    <string name="text_settings">Настройки темы</string>
+    <string name="button_back">Назад</string>
 </resources>
