+++ app/src/main/res/values/themes.xml
@@ -22,6 +22,7 @@
         <item name="colorAccent">@color/black</item>
 
     </style>
+
     <style name="MyThemeGrey" parent="Theme.MaterialComponents.NoActionBar">
         <item name="colorPrimary">@color/grey</item>
         <item name="colorPrimaryVariant">@color/grey</item>
@@ -29,6 +30,15 @@
         <item name="colorPrimaryDark">@color/grey</item>
         <item name="colorAccent">@color/black</item>
     </style>
+
+    <style name="ThemeCalculator.NoActionBar">
+        <item name="windowActionBar">false</item>
+        <item name="windowNoTitle">true</item>
+    </style>
+
+    <style name="ThemeCalculator.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
+
+    <style name="ThemeCalculator.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
 
 
 </resources>
